Created Play method. Now all the question instances are created in the ComputerGeekGame constructor.
Created a list for each of the types of questions as well as methods which randomize the order or the questions every time Play() is called.

The score and time are both initiated at the beginning of Play() and they are calculate and printed to the screen when the game is over.

The method in charge of randomizing the answer options is now up and running and seems to be working great.

Changed some small formatting issues that make the game look better overall.